requirements.txt
Django==1.11.6
django-filter==1.0.4
djangorestframework==3.7.0
pymongo==3.5.1
python==3.6.2